+++ contracts/canonical/CrunaGuardian.sol
@@ -14,6 +14,8 @@ import {FlexiTimelockController} from "../utils/FlexiTimelockController.sol";
  * - manager to trust a new plugin implementation and allow managed transfers
  */
 contract CrunaGuardian is ICrunaGuardian, IVersioned, FlexiTimelockController {
+  bool private _allowUntrusted;
+
   /**
    * @notice Error returned when the arguments are invalid
    */
@@ -60,7 +62,18 @@ contract CrunaGuardian is ICrunaGuardian, IVersioned, FlexiTimelockController {
     emit TrustedImplementationUpdated(nameId, implementation, trusted);
   }
 
+  /// @dev see {ICrunaGuardian-trustedImplementation}
   function trustedImplementation(bytes4 nameId, address implementation) external view override returns (bool) {
     return _trustedImplementations[bytes32(nameId) | bytes32(uint256(uint160(implementation)))];
+  }
+
+  /// @dev see {ICrunaGuardian-allowUntrusted}
+  function allowUntrusted(bool allowUntrusted_) external onlyRoleOrOpenRole(DEFAULT_ADMIN_ROLE) {
+    _allowUntrusted = allowUntrusted_;
+  }
+
+  /// @dev see {ICrunaGuardian-allowingUntrusted}
+  function allowingUntrusted() external view returns (bool) {
+    return _allowUntrusted;
   }
 }