+++ contracts/manager/CrunaManager.sol
@@ -213,7 +213,7 @@ contract CrunaManager is Actor, CrunaManagerBase {
     bool trusted = Canonical.crunaGuardian().trustedImplementation(nameId_, pluginProxy);
     if (!trusted)
       if (canManageTransfer)
-        if (!_vault().allowUntrustedTransfers()) {
+        if (!Canonical.crunaGuardian().allowingUntrusted()) {
           revert UntrustedImplementationsNotAllowedToMakeTransfers();
         }
     _preValidateAndCheckSignature(
@@ -352,7 +352,7 @@ contract CrunaManager is Actor, CrunaManagerBase {
     _removeLockIfExpired(pluginNameId, salt);
     if (!_pluginByKey[_key].canManageTransfer) revert PluginNotAuthorizedToManageTransfer();
     if (!_pluginByKey[_key].trusted)
-      if (!_vault().allowUntrustedTransfers()) revert UntrustedImplementationsNotAllowedToMakeTransfers();
+      if (!Canonical.crunaGuardian().allowingUntrusted()) revert UntrustedImplementationsNotAllowedToMakeTransfers();
     _resetOnTransfer(pluginNameId, salt);
     // In theory, the vault may revert, blocking the entire process
     // We allow it, assuming that the vault implementation has the
@@ -497,7 +497,7 @@ contract CrunaManager is Actor, CrunaManagerBase {
     uint256 timeLock
   ) internal virtual {
     if (!_pluginByKey[_key].trusted)
-      if (!_vault().allowUntrustedTransfers()) revert UntrustedImplementationsNotAllowedToMakeTransfers();
+      if (!Canonical.crunaGuardian().allowingUntrusted()) revert UntrustedImplementationsNotAllowedToMakeTransfers();
     CrunaPluginBase plugin_ = _plugin(nameId_, salt);
     if (!plugin_.requiresToManageTransfer()) revert NotATransferPlugin();
     if (change == PluginChange.Authorize) {